easyocr/app/requirements.txt
fastapi
uvicorn
easyocr
python-multipart
pillow
numpy
